remove or change some profile actions and data
